% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessProperty.R
\name{accessProperty}
\alias{accessProperty}
\title{Access Properties}
\usage{
accessProperty(
  api.key,
  ontology,
  id = NA_character_,
  sub = NA_character_,
  output.mode = "content"
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any
API call. It is used within {cedarr} as a header for http
requests. An API key is linked to a CEDAR account
(https://cedar.metadatacenter.org/profile)}

\item{ontology}{character. Name of the ontology to query properties from.}

\item{id}{character. Either NA to list properties, "roots" to get the root
properties of the ontology or a property ID (at URL format), to find a
precise property.}

\item{sub}{character. Either NA to get the precise property alone, or one
among "tree", "children", "descendants", "parents" to get related properties
according to this value. Not evaluated if \code{id} is set to NA or "roots".}

\item{output.mode}{character. "full" will return the whole
response object (from {httr}) or "content" will fetch the
interest values from the response object. Getting the whole
object might be interesting to have a look at system metadata,
or in case of error to debug the connection. (defaults to
"content")}
}
\value{
If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Function to access the CEDAR properties suite.
}
\details{
This function matches the following queries from the Swagger UI
(https://terminology.metadatacenter.org/api/#/):

\itemize {
\item {\verb{/ontologies/\{ontology\}/properties}}
\item {\verb{/ontologies/\{ontology\}/properties/roots}}
\item {\verb{/ontologies/\{ontology\}/properties/\{id\}}}
\item {\verb{/ontologies/\{ontology\}/properties/\{id\}/tree}}
\item {\verb{/ontologies/\{ontology\}/properties/\{id\}/children}}
\item {\verb{/ontologies/\{ontology\}/properties/\{id\}/descendants}}
\item {\verb{/ontologies/\{ontology\}/properties/\{id\}/parents}}
}

These differents requests are differenciated by the \code{id} and \code{sub}
arguments.
}
\examples{
my.api.key <- readline()

# Query 1: list the roots properties in ENVO

result1 <- cedarr::accessProperty(
  my.api.key,
  "ENVO",
  id = "roots",
  sub = "smurf" # ignored
)

View(result1)

# Query 2: get the parents properties for "alternative term" in ENVO

result2 <- cedarr::accessProperty(
  my.api.key,
  "ENVO",
  id = "http://purl.obolibrary.org/obo/IAO_0000118",
  sub = "parents"
)

View(result2)


}
