% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessValues.R
\name{accessValues}
\alias{accessValues}
\title{Explore values in CEDAR}
\usage{
accessValues(
  api.key,
  vs.collection,
  id,
  sub = NA_character_,
  output.mode = "content",
  page.index = 1,
  page.size = 50
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any API call. It
is used within {cedarr} as a header for http requests. An API key is linked
to a CEDAR account (https://cedar.metadatacenter.org/profile)}

\item{vs.collection}{character. A VS collection id.}

\item{id}{character. A value ID in a value sets from the VS collection.}

\item{sub}{character. A sub-item to fetch from the value. Not
evaluated if \code{id} is set to NA. Else, can be NA, "value-set", "tree"
or "all-values". "value-set" allows to get the value-set the value
belongs to.}

\item{output.mode}{character. "full" will return the whole
response object (from {httr}) or "content" will fetch the
interest values from the response object. Getting the whole
object might be interesting to have a look at system metadata,
or in case of error to debug the connection. (defaults to
"content")}

\item{page.index}{integer. Index of the page to be returned (defaults to 1st page).}

\item{page.size}{integer. Number of results per page, capped at 50. (defaults
to 50).}
}
\value{
\itemize{
\item{If a single value ID is provided: an entry for the corresponding value.}
\item{If a value ID is provided and \code{sub} = "tree": get value tree (only for
regular values).}
\item{If a value ID is provided and \code{sub} = "all-values": find all values
in the value set that the given value belongs to.}
\item{If a value ID is provided and \code{sub} = "value-set": an entry for the
value set the target values belong to.}
}

If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Find value by id.
}
\details{
This function matches the following queries from the Swagger UI
(https://terminology.metadatacenter.org/api/#/):

\itemize {
\item {\verb{/vs-collections/\{vs_collection\}/values/\{id\}}}
\item {\verb{/vs-collections/\{vs_collection\}/values/\{id\}/tree}}
\item {\verb{/vs-collections/\{vs_collection\}/values/\{id\}/all-values}}
\item {\verb{/vs-collections/\{vs_collection\}/values/\{id\}/value-set}}
}

These differents requests are differenciated by the \code{id} and \code{sub}
arguments.
}
\examples{
my.api.key <- readline()

result <- cedarr::accessValues(
  my.api.key,
  vs.collection = "CEDARVS",
  id = ??? # WIP
)

View(result)

}
