% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessVSCollections.R
\name{accessValueSets}
\alias{accessValueSets}
\title{Search values in CEDAR}
\usage{
accessValueSets(
  api.key,
  vs.collection,
  id = NA_character_,
  sub = NA_character_,
  output.mode = "content",
  page = 1,
  page.size = 50
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any API call. It is used within {cedarr}
as a header for http requests. An API key is linked to a CEDAR account (https://cedar.metadatacenter.org/profile)}

\item{output.mode}{character. "full" will return the whole response object (from {httr}) or "content" will
fetch the interest values from the response object. Getting the whole object might be interesting to
have a look at system metadata, or in case of error to debug the connection. (defaults to "content")}
}
\value{
If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Find all value set collections.
}
\examples{
my.api.key <- readline()

result <- cedarr::accessValueSets(
  my.api.key,
  vs.collection = "CEDARVS",
  id = "http://www.semanticweb.org/jgraybeal/ontologies/2015/7/cedarvaluesets#Study_File_Type"
)

View(result)

}
