% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedar.client-class.R
\docType{class}
\name{cedar.client-class}
\alias{cedar.client-class}
\alias{cedar.client}
\title{CEDAR API Access client turned into an R class.}
\description{
CEDAR API Access client turned into an R class.
}
\section{Slots}{

\describe{
\item{\code{apikey}}{An API Key is required to access any API call. It is used within BORAA methods as
a header for http requests. An API key is linked to a CEDAR account (\url{cedar.metadatacenter.org}).}
}}

