% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cedar.get}
\alias{cedar.get}
\title{CEDAR HTTP get method}
\usage{
cedar.get(api.key, url, ..., output.mode = "content")
}
\arguments{
\item{api.key}{character. An API Key is required to access any
API call. It is used within {cedarr} as a header for http
requests. An API key is linked to a CEDAR account
(https://cedar.metadatacenter.org/profile)}

\item{url}{character. An URL to the CEDAR resource to query data from.}

\item{...}{arguments to input in the request body.}

\item{output.mode}{character. "full" will return the whole
response object (from {httr}) or "content" will fetch the
interest values from the response object. Getting the whole
object might be interesting to have a look at system metadata,
or in case of error to debug the connection. (defaults to
"content")}
}
\value{
If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Internal function wrapping the GET method.
}
