% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkLength}
\alias{checkLength}
\title{Correct arguments length.}
\usage{
checkLength(..., check, env = .GlobalEnv)
}
\arguments{
\item{...}{list. A list of arguments to check for length.}

\item{check}{ArgCheck. An ArgCheck object from package {ArgumentCheck}.}

\item{env}{environment. The environment in which  ... is being evaluated.}
}
\value{
The \code{check} argument.

Check for each given argument in \code{...} that their length > 1.
}
\description{
Correct arguments length.
}
\details{
This function will perform side-effects actions, shortening excessively long
items to length-one items.
}
