% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constantCheck}
\alias{constantCheck}
\title{Constant checks}
\usage{
constantCheck(..., check, env = .GlobalEnv)
}
\arguments{
\item{...}{list. A list of arguments to check for validity.}

\item{check}{ArgCheck. An ArgCheck object from package {ArgumentCheck}.}

\item{env}{environment. The environment in which args is being evaluated.}
}
\description{
These checks are found in every function of {cedarr}.
}
