% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listValueSets.R
\name{listValueSets}
\alias{listValueSets}
\alias{listVSCollections}
\title{List value sets and value sets collections}
\usage{
listValueSets(api.key, output.mode = "content")

listVSCollections(api.key, output.mode = "content")
}
\arguments{
\item{api.key}{character. An API Key is required to access any
API call. It is used within {cedarr} as a header for http
requests. An API key is linked to a CEDAR account
(https://cedar.metadatacenter.org/profile)}

\item{output.mode}{character. "full" will return the whole
response object (from {httr}) or "content" will fetch the
interest values from the response object. Getting the whole
object might be interesting to have a look at system metadata,
or in case of error to debug the connection. (defaults to
"content")}
}
\value{
An exhaustive list of value sets for any value set collection.

An exhaustive list of the VS collections in CEDAR.

If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Find all value sets or value sets collections.
}
\section{Functions}{
\itemize{
\item \code{listVSCollections}: 
}}

\examples{
\dontrun{
my.api.key <- readline()

result1 <- cedarr::listValueSets(
  my.api.key
)

View(result1)
}

\dontrun{
my.api.key <- readline()

result2 <- cedarr::listVSCollections(
  my.api.key,
)

View(result2)
}

}
