% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessOntology.R
\name{accessOntology}
\alias{accessOntology}
\title{Search values in CEDAR}
\usage{
accessOntology(
  api.key,
  ontology,
  output.mode = "content",
  page = 1,
  page.size = 50
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any API call. It is used within {cedarr}
as a header for http requests. An API key is linked to a CEDAR account (https://cedar.metadatacenter.org/profile)}

\item{ontology}{character. Ontology name to display.}

\item{output.mode}{character. "full" will return the whole response object (from {httr}) or "content" will
fetch the interest values from the response object. Getting the whole object might be interesting to
have a look at system metadata, or in case of error to debug the connection. (defaults to "content")}

\item{page}{integer. Index of the page to be returned (defaults to 1st page).}

\item{page.size}{integer. Number of results per page, capped at 50. (defaults to 50).}
}
\description{
Get all classes from a specific ontology (including both regular and provisional classes).
}
