% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessProperty.R
\name{accessProperty}
\alias{accessProperty}
\title{Access properties}
\usage{
accessProperty(
  api.key,
  ontology,
  id = NA_character_,
  sub = NA_character_,
  output.mode = "content"
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any
API call. It is used within {cedarr} as a header for http
requests. An API key is linked to a CEDAR account
(https://cedar.metadatacenter.org/profile)}

\item{ontology}{character. Name of the ontology to query properties from.}

\item{id}{character. Either NA to list properties, "roots" to get the root
properties of the ontology or a property ID (at URL format), to find a
precise property.}

\item{sub}{character. Either NA to get the precise property alone, or one
among "tree", "children", "descendants", "parents" to get related properties
according to this value. Ignored if \code{id} is set to NA or "roots".}
}
\value{
If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Function to access the CEDAR properties suite.
}
