% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessValueSets.R
\name{accessValueSets}
\alias{accessValueSets}
\title{Access Value Sets}
\usage{
accessValueSets(
  api.key,
  vs.collection,
  id = NA_character_,
  sub = NA_character_,
  output.mode = "content",
  page.index = 1,
  page.size = 50
)
}
\arguments{
\item{api.key}{character. An API Key is required to access any
API call. It is used within {cedarr} as a header for http
requests. An API key is linked to a CEDAR account
(https://cedar.metadatacenter.org/profile)}

\item{vs.collection}{character. A VS collection id.}

\item{id}{character. A value ID in the VS collection at URL format.
Not providing it (setting it to NA) will list all value sets in the \code{vs.collection}.}

\item{sub}{character. A sub-item to fetch from the value: can be NA, "tree",
or "values". Not evaluated if \code{id} is set to NA.}

\item{output.mode}{character. "full" will return the whole
response object (from {httr}) or "content" will fetch the
interest values from the response object. Getting the whole
object might be interesting to have a look at system metadata,
or in case of error to debug the connection. (defaults to "content")}

\item{page.index}{integer. Index of the page to be returned (defaults to 1st page).}

\item{page.size}{integer. Number of results per page, capped at 50. (defaults
to 50).}
}
\value{
\itemize {
\item {If only a VS collection ID is provided: a list of the value sets in
this VS collection.}
\item {If a VS collection ID and a value set ID are provided: an entry for
the corresponding value set. With \code{sub} = "tree", retrieve the root values.
If \code{sub} = "values", retrieve the value set's values.}
}

If \code{output.mode = "full"}, the whole http response object (see httr::response).
It is structured as a list with response metadata wrapping the \code{content} item
which contains the wanted result.

If \code{output.mode = "content"}, the \code{content} item is directly returned, containing
database metadata and the interesting information in the \code{collection} subitem.
}
\description{
Get access to all or part of the value sets by VS collection.
}
\details{
This function matches the following queries from the Swagger UI
(https://terminology.metadatacenter.org/api/#/):

\itemize {
\item {\verb{/vs-collection/\{vs_collection\}/value-sets}}
\item {\verb{/vs-collection/\{vs_collection\}/value-sets/\{id\}}}
\item {\verb{/vs-collection/\{vs_collection\}/value-sets/\{id\}/tree}}
\item {\verb{/vs-collection/\{vs_collection\}/value-sets/\{id\}/values}}
}

These differents requests are differenciated by the \code{id} and \code{sub}
arguments.
}
\examples{
my.api.key <- readline()

result <- cedarr::accessValueSets(
  my.api.key,
  vs.collection = "CEDARVS",
  id = "http://www.semanticweb.org/jgraybeal/ontologies/2015/7/cedarvaluesets#Study_File_Type"
)

View(result)

}
