% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listValueSets.R
\name{listValueSets}
\alias{listValueSets}
\title{Search values in CEDAR}
\usage{
listValueSets(api.key, output.mode = "content")
}
\arguments{
\item{api.key}{character. An API Key is required to access any API call. It is used within {cedarr}
as a header for http requests. An API key is linked to a CEDAR account (https://cedar.metadatacenter.org/profile)}

\item{output.mode}{character. "full" will return the whole response object (from {httr}) or "content" will
fetch the interest values from the response object. Getting the whole object might be interesting to
have a look at system metadata, or in case of error to debug the connection. (defaults to "content")}
}
\description{
Find all value sets.
}
